Do not explicitly load the worker MPM
Several reasons for this:

1. Unlike most Apache modules used by webvm, mpm_worker used symlinks
to /etc/apache2/mods-available/. This path is specific to Debian-like
distributions;

2. On modern systems (i.e. ones which support thread-safe polling)
it is better to use mpm_event, which is more resource-efficient
mod_perl works fine with both of them.

3. Assuming we do not want to choose mpm_worker over mpm_event, there
is no need to explicitly choose an MPM - mod_perl will choose the right
one depending on whether it has been built with or without thread
support.