Do not force-load mod_perl
Fun fact: we do not in fact use mod_perl. All the server-side code is
executed as CGI. Indeed, attempting to enable mod_perl even in
maximum-compatibility mode (i.e. using ModPerl::PerlRun) does not work -
the first error that shows up is Otter code being unable to extract
version identifier from the script path because mod_perl seems to handle
this differently than plain CGI, and it is quite conceivable that
addressing this would reveal additional errors due to complexity of our
code.

Besides, Apache typically auto-loads mod_perl so even if we used it it
wouldn't be necessary to load it by hand. On the other hand, Apache
fails to start if it cannot load a module that has been explicitly
requested.

The upshot: with this change in place Otter server no longer depends on
mod_perl.