If no Unix MPM loaded automatically, try loading mpm_event
Turns out that at least on some versions of Ubuntu the event MPM is
*not* loaded automatically, causing a start-up error. Therefore, tell
Apache to explicity attempt loading the event MPM unless one of the
common Unix MPMs has been found.

Note that this solution is somewhat fragile because it will still
attempt to load mpm_event:
 - on systems with some MPM other than event, prefork or worker already
   loaded; or
 - on systems with no MPM loaded but which do not support mpm_event.
Both of these conditions essentially cover the same scenario: running on
a non-Unix OS. Shouldn't be a problem for us given the Otter server is
to always run on Linux but it is worth keeping in mind.